+++ packages/falcon-server-env/src/models/Extension.ts
@@ -162,6 +162,9 @@ export default abstract class Extension {
             : typeDefs
                 // Removing "extend type X" to avoid "X type missing" errors
                 .replace(/extend\s+type/gm, 'type')
+                // Removing directives
+                .replace(/@(\w+)\(.*\)/gm, '')
+                .replace(/@(\w+)/gm, '')
                 // Removing type references from the base schema types
                 .replace(/:\s*(\w+)/gm, ': Int')
                 .replace(/\[\s*(\w+)\s*]/gm, '[Int]')