+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 ## Falcon (Next)
 
-### Falcon E-commerce UI Kit v0.4.0 (Next)
+### Falcon E-commerce UI Kit (Next)
 
 - added basic component set (queries, mutations and ui) to build filters ([#365](https://github.com/deity-io/falcon/pull/365))