+++ CHANGELOG.md
@@ -13,7 +13,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 ### Falcon Shop Extension (Next)
 
-- `menu`, `Product` and `Product.breadcrumbs` resolvers are cached now ([#374](https://github.com/deity-io/falcon/pull/374))
+- `menu`, `Category`, `Category.products`, `Product` and `Product.breadcrumbs` resolvers are cached now ([#374](https://github.com/deity-io/falcon/pull/374))
 
 ### Falcon E-commerce UI Kit (Next)