+++ CHANGELOG.md
@@ -6,7 +6,22 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 ## Falcon (Next)
 
+### Falcon Magento2 API (next)
+
+- fixed issue with merging guest's with customer's carts ([#394](https://github.com/deity-io/falcon/pull/394))
+- fixed issue with 0 price for the first product added to cart by authorized user ([#394](https://github.com/deity-io/falcon/pull/394))
+
+### Falcon-Server (Next)
+
+- introduced `@cache` GraphQL directive to cache resolver results ([#374](https://github.com/deity-io/falcon/pull/374))
+- `url` resolver is cached now ([#374](https://github.com/deity-io/falcon/pull/374))
+
+### Falcon Shop Extension (Next)
+
+- `menu`, `Category`, `Category.products`, `Product` and `Product.breadcrumbs` resolvers are cached now ([#374](https://github.com/deity-io/falcon/pull/374))
+
 ### Falcon Client (Next)
+
 - fixed absolute path to output directory embedded in bundle ([385](https://github.com/deity-io/falcon/issues/385))
 
 ### Falcon E-commerce UI Kit (Next)