+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Versions marked with a number and date (e.g. Falcon Client v0.1.0 (2018-10-05)) 
 
 ## Falcon vNEXT
 
-### Falcon Pkg v0.0.1
+### Falcon Scripts v0.0.1
 
 - initial release